split file structure
